+++ evennia/utils/eveditor.py
@@ -801,7 +801,7 @@ class EvEditor(object):
         try:
             self._buffer = self._loadfunc(self._caller)
             if not isinstance(self._buffer, str):
-                self._buffer = to_str(self._buffer, force_string=True)
+                self._buffer = to_str(self._buffer)
                 self._caller.msg("|rNote: input buffer was converted to a string.|n")
         except Exception as e:
             from evennia.utils import logger